+++ Sources/CGFloat+Joystick.swift
@@ -1,0 +1,14 @@
+//
+//  CGFloat+Joystick.swift
+//
+//
+//  Created by tichise on 2024/01/01.
+//
+
+import Foundation
+
+extension CGFloat {
+    func text() -> String {
+        return String(format: "%.00f", Float(self))
+    }
+}
