+++ Sources/JoyStickState.swift
@@ -1,0 +1,16 @@
+//
+//  JoyStickState.swift
+//  
+//
+//  Created by tichise on 2024/01/01.
+//
+
+import Foundation
+
+public enum JoyStickState: String {
+    case up
+    case down
+    case left
+    case right
+    case center
+}
