+++ Sources/OMJoyStick.swift
@@ -8,115 +8,6 @@
 
 import SwiftUI
 
-struct SmallRing: View {
-    var smallRingDiameter: CGFloat
-    var subRingColor: Color
-    
-    var body: some View {
-        Circle().fill(subRingColor)
-            .frame(width: smallRingDiameter, height: smallRingDiameter)
-    }
-}
-
-struct BigRing: View {
-    @Environment(\.colorScheme) var colorScheme
-    
-    var bigRingNormalBackgroundColor: Color
-    var bigRingDarkBackgroundColor: Color
-    var bigRingStrokeColor: Color
-    
-    var bigRingDiameter: CGFloat
-    
-    // 8等分のラインを表示するか制御する
-    @Binding var isSplitLine: Bool
-    
-    var body: some View {
-        ZStack {
-            Circle().stroke(bigRingStrokeColor, lineWidth: 10)
-                .frame(width: bigRingDiameter, height: bigRingDiameter)
-            
-            Circle().fill(colorScheme == .dark ? bigRingDarkBackgroundColor : bigRingNormalBackgroundColor)
-                .frame(width: bigRingDiameter, height: bigRingDiameter)
-            
-            if isSplitLine {
-                let radius = min(bigRingDiameter, bigRingDiameter) / 2
-                let center = CGPoint(x: bigRingDiameter / 2, y: bigRingDiameter / 2)
-                
-                 
-                    ForEach(0..<8) { i in
-                        Path { path in
-                            path.move(to: center)
-                            let angle = CGFloat(i) * .pi / 4 // 45度をラジアンに変換
-                            let endX = center.x + radius * cos(angle)
-                            let endY = center.y + radius * sin(angle)
-                            path.addLine(to: CGPoint(x: endX, y: endY))
-                        }
-                        .stroke(bigRingStrokeColor, lineWidth: 1)
-                    }
-            }
-            
-            
-
-        }.frame(width: bigRingDiameter, height: bigRingDiameter)
-    }
-}
-
-// BingRingのPreview用
-struct BigRing_Previews: PreviewProvider {
-    static var previews: some View {
-        BigRing(bigRingNormalBackgroundColor: .white, bigRingDarkBackgroundColor: .black, bigRingStrokeColor: .primary, bigRingDiameter: 100, isSplitLine: .constant(false))
-    }
-}
-
-extension CGFloat {
-    func text() -> String {
-        return String(format: "%.00f", Float(self))
-    }
-}
-
-public enum JoyStickState: String {
-    case up
-    case down
-    case left
-    case right
-    case center
-}
-
-public struct IconSetting {
-    var leftIcon: Image
-    var rightIcon: Image
-    var upIcon: Image
-    var downIcon: Image
-    
-    public init(leftIcon: Image, rightIcon: Image, upIcon: Image, downIcon: Image) {
-        self.leftIcon = leftIcon
-        self.rightIcon = rightIcon
-        self.upIcon = upIcon
-        self.downIcon = downIcon
-    }
-}
-
-public struct ColorSetting {
-    var subRingColor: Color
-    var bigRingNormalBackgroundColor: Color
-    var bigRingDarkBackgroundColor: Color
-    var bigRingStrokeColor: Color
-    var iconColor: Color
-    
-    public init(subRingColor: Color = .primary,
-    bigRingNormalBackgroundColor: Color = .white,
-    bigRingDarkBackgroundColor: Color = .black,
-    bigRingStrokeColor: Color = Color.primary,
-    iconColor: Color = Color.primary) {
-        
-        self.subRingColor = subRingColor
-        self.bigRingNormalBackgroundColor = bigRingNormalBackgroundColor
-        self.bigRingDarkBackgroundColor = bigRingDarkBackgroundColor
-        self.bigRingStrokeColor = bigRingStrokeColor
-        self.iconColor = iconColor
-    }
-}
-
 public struct OMJoystick: View {
     
     var iconColor: Color
